Move ring related configuration to the Ring drawable
The playfield shouldn't bear that responsibility in the first place
